# 02-go-web/middlewares/documents/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Movies API.",
        "title": "Movies API",
        "contact": {
            "name": "API Support",
            "email": "random@gmail.com"
        },
        "version": "1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/movies": {
            "get": {
                "description": "Get all movies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get movies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseBodyGetMovies"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseBodyGetMovies"
                        }
                    }
                }
            }
        },
        "/movies/{title}": {
            "get": {
                "description": "Get a movie by title",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get movie by title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie title",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseBodyGetMovieByTitle"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseBodyGetMovieByTitle"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseBodyGetMovieByTitle"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.MovieJSON": {
            "type": "object",
            "properties": {
                "genre": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "handlers.ResponseBodyGetMovieByTitle": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.MovieJSON"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.ResponseBodyGetMovies": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.MovieJSON"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}